move PathPrefix interface in types
