+++ web/ui/react-app/src/pages/alerts/Alerts.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import { RouteComponentProps } from '@reach/router';
-import PathPrefixProps from '../../PathPrefixProps';
+import PathPrefixProps from '../../types/PathPrefixProps';
 import { useFetch } from '../../hooks/useFetch';
 import { withStatusIndicator } from '../../common/withStatusIndicator';
 import AlertsContent, { RuleStatus, AlertsProps } from './AlertContents';
