+++ web/ui/react-app/src/pages/graph/PanelList.tsx
@@ -6,7 +6,7 @@ import { Alert, Button, Col, Row } from 'reactstrap';
 import Panel, { PanelOptions, PanelDefaultOptions } from './Panel';
 import { decodePanelOptionsFromQueryString, encodePanelOptionsToQueryString } from '../../utils/urlParams';
 import Checkbox from '../../common/Checkbox';
-import PathPrefixProps from '../../PathPrefixProps';
+import PathPrefixProps from '../../types/PathPrefixProps';
 import { generateID } from '../../utils/func';
 
 export type MetricGroup = { title: string; items: string[] };
