+++ web/ui/react-app/src/pages/tsdbStatus/TSDBStatus.tsx
@@ -4,7 +4,7 @@ import { Alert, Table } from 'reactstrap';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faSpinner } from '@fortawesome/free-solid-svg-icons';
 import { useFetch } from '../../hooks/useFetch';
-import PathPrefixProps from '../../PathPrefixProps';
+import PathPrefixProps from '../../types/PathPrefixProps';
 
 export interface Stats {
   name: string;
